Feat : thumbnail 컴포넌트 분리 및 alt 프로퍼티 추가
